// src/types/Item.ts
export type item = {
    inventory: number;
    price?: number | undefined;
    title?: string | '';
    image?: string | "";
}

// export type image = {
//     assetId?: any;
//     base64?: string;
//     cancelled: boolean;
//     exif?: any,
//     width: number;
//     height: number;
//     type: string;
//     uri: string;
// }